# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: openleash API
  description: |
    Local-first authorization and cryptographic proof sidecar for AI agents.

    OpenLeash evaluates YAML-based policies against agent action requests and,
    when an action is allowed, issues PASETO v4.public proof tokens that
    counterparties can verify offline or online.

    ## Authentication

    The API uses four security zones:

    - **Public** — No authentication. Health checks, public key distribution,
      proof verification, and the playground.
    - **Agent** — Ed25519 request signing via five headers (`X-Agent-Id`,
      `X-Timestamp`, `X-Nonce`, `X-Body-Sha256`, `X-Signature`). Used for
      `/v1/authorize` and `/v1/agent/*` endpoints.
    - **Owner** — PASETO v4.public session token via `Authorization: Bearer`
      header (or `openleash_session` cookie for GUI). Issued on login.
      Used for all `/v1/owner/*` endpoints.
    - **Admin** — Bearer token and/or localhost check depending on the
      `admin.mode` setting (`localhost`, `token`, or `localhost_or_token`).
      Used for all `/v1/admin/*` endpoints.

    ## Typical flow

    1. An admin creates an **owner** via the Admin API and generates a
       **setup invite**.
    2. The owner completes setup (`POST /v1/owner/setup`) with the invite
       token and a passphrase, then logs in (`POST /v1/owner/login`).
    3. The owner creates **policies** via the Owner API.
    4. An agent calls `/v1/agents/registration-challenge` then
...
servers:
...
paths:
...
  /v1/owner/setup:
    post:
      summary: Complete owner setup
      operationId: ownerSetup
      tags: [Owner]
      description: |
        Completes owner account setup using a one-time setup invite
        (generated by an admin). Sets the owner's passphrase for future
        login. The invite token is verified against its stored hash and
        marked as used.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [invite_id, invite_token, passphrase]
              properties:
                invite_id:
                  type: string
                  format: uuid
                  description: ID of the setup invite.
                invite_token:
                  type: string
                  description: The plaintext invite token (received from admin).
                passphrase:
                  type: string
                  description: The passphrase to set (minimum 8 characters).
      responses:
        "200":
          description: Setup complete.
          content:
            application/json:
              schema:
                type: object
                required: [status, owner_principal_id]
                properties:
                  status:
                    type: string
                    example: setup_complete
                  owner_principal_id:
                    type: string
                    format: uuid
        "400":
          description: Invalid request, invite used, invite expired, or weak passphrase.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/owner/login:
    post:
      summary: Owner login
      operationId: ownerLogin
      tags: [Owner]
      description: |
        Authenticates an owner with their principal ID and passphrase.
        Returns a PASETO v4.public session token valid for the configured
        TTL (default 8 hours). The session is stateless — the token
        contains the owner's identity and expiry.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [owner_principal_id, passphrase]
              properties:
                owner_principal_id:
                  type: string
                  format: uuid
                passphrase:
                  type: string
      responses:
        "200":
          description: Login successful.
          content:
            application/json:
              schema:
                type: object
                required: [token, expires_at, owner_principal_id]
                properties:
                  token:
                    type: string
                    description: PASETO v4.public session token (starts with `v4.public.`).
                  expires_at:
                    type: string
                    format: date-time
                  owner_principal_id:
                    type: string
                    format: uuid
        "401":
          description: Invalid credentials or setup not completed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/owner/logout:
    post:
      summary: Owner logout
      operationId: ownerLogout
      tags: [Owner]
      description: |
        Logs out the owner. Since sessions are stateless (no server-side
        session store), this is primarily an audit event. Clients should
        discard the session token.
      security:
        - OwnerSession: []
      responses:
        "200":
          description: Logged out.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    example: logged_out
  /v1/owner/profile:
    get:
      summary: Get owner profile
      operationId: getOwnerProfile
      tags: [Owner]
      description: |
        Returns the authenticated owner's profile. Sensitive fields
        (passphrase hash/salt) are stripped from the response.
      security:
        - OwnerSession: []
      responses:
        "200":
          description: Owner profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Owner"
    put:
      summary: Update owner profile
      operationId: updateOwnerProfile
      tags: [Owner]
      description: |
        Updates the authenticated owner's profile fields including
        display name, contact identities, government IDs, company IDs,
        signatories, and signatory rules. Identity validation is applied
        and assurance level is recomputed.
      security:
        - OwnerSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name:
                  type: string
                contact_identities:
                  type: array
                  items:
                    $ref: "#/components/schemas/ContactIdentity"
                government_ids:
                  type: array
                  items:
                    $ref: "#/components/schemas/GovernmentId"
                company_ids:
                  type: array
                  items:
                    $ref: "#/components/schemas/CompanyId"
                signatories:
                  type: array
                  items:
                    $ref: "#/components/schemas/Signatory"
                signatory_rules:
                  type: array
                  items:
                    $ref: "#/components/schemas/SignatoryRule"
      responses:
        "200":
          description: Updated owner profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Owner"
  /v1/owner/agents:
    get:
      summary: List owner's agents
      operationId: listOwnerAgents
      tags: [Owner]
      description: |
        Returns all agents belonging to the authenticated owner.
      security:
        - OwnerSession: []
      responses:
        "200":
          description: Agent list.
          content:
            application/json:
              schema:
                type: object
                required: [agents]
                properties:
                  agents:
                    type: array
                    items:
                      $ref: "#/components/schemas/AgentRegistration"
  /v1/owner/agents/{agentId}:
    put:
      summary: Update agent status
      operationId: updateOwnerAgent
      tags: [Owner]
      description: |
        Updates an agent's status (e.g., revoke). Only agents belonging
        to the authenticated owner can be updated. Cross-owner access
        returns 404.
      security:
        - OwnerSession: []
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The agent's `agent_principal_id`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [ACTIVE, REVOKED]
      responses:
        "200":
          description: Agent updated.
          content:
            application/json:
              schema:
                type: object
                required: [agent_principal_id, agent_id, status]
                properties:
                  agent_principal_id:
                    type: string
                    format: uuid
                  agent_id:
                    type: string
                  status:
                    type: string
        "404":
          description: Agent not found or not owned by this owner.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/owner/policies:
    get:
      summary: List owner's policies
      operationId: listOwnerPolicies
      tags: [Owner]
      description: |
        Returns all policies belonging to the authenticated owner,
        including their YAML content.
      security:
        - OwnerSession: []
      responses:
        "200":
          description: Policy list.
          content:
            application/json:
              schema:
                type: object
                required: [policies]
                properties:
                  policies:
                    type: array
                    items:
                      $ref: "#/components/schemas/PolicyEntry"
    post:
      summary: Create policy
      operationId: createOwnerPolicy
      tags: [Owner]
      description: |
        Creates a new policy owned by the authenticated owner. The YAML
        is validated against the policy schema. Optionally binds the
        policy to a specific agent.
      security:
        - OwnerSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [policy_yaml]
              properties:
                policy_yaml:
                  type: string
                  description: The policy YAML content.
                applies_to_agent_principal_id:
                  type: string
                  format: uuid
                  nullable: true
                  description: If set, binds this policy to a specific agent.
      responses:
        "200":
          description: Policy created.
          content:
            application/json:
              schema:
                type: object
                required: [policy_id, owner_principal_id]
                properties:
                  policy_id:
                    type: string
                    format: uuid
                  owner_principal_id:
                    type: string
                    format: uuid
                  applies_to_agent_principal_id:
                    type: string
                    format: uuid
                    nullable: true
...
  /v1/owner/approval-requests:
    get:
      summary: List approval requests
      operationId: listOwnerApprovalRequests
      tags: [Owner]
      description: |
        Returns approval requests for the authenticated owner. Supports
        filtering by status via query parameter.
      security:
        - OwnerSession: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, APPROVED, DENIED, EXPIRED]
          description: Filter by approval request status.
      responses:
        "200":
          description: Approval request list.
          content:
            application/json:
              schema:
                type: object
                required: [approval_requests]
                properties:
                  approval_requests:
                    type: array
                    items:
                      $ref: "#/components/schemas/ApprovalRequestOwnerView"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        Admin token configured in `config.yaml` under `admin.token`.
        Required when `admin.mode` is `token` or for remote access in
        `localhost_or_token` mode.

    OwnerSession:
      type: http
      scheme: bearer
      description: |
        PASETO v4.public session token issued by `POST /v1/owner/login`.
        Contains `purpose: 'owner_session'` claim to distinguish from
        proof tokens. Default TTL is 8 hours (configurable via
        `sessions.ttl_seconds` in `config.yaml`).
...